+++ tools/c7n_gcp/tests/test_compute.py
@@ -571,20 +571,3 @@ class ProjectTest(BaseTest):
 
         self.assertEqual(len(resources), 1)
         self.assertEqual(resources[0]['name'], 'gcp-lab-custodian')
-
-
-class TestDataprocCluster(BaseTest):
-
-    def test_query(self):
-        project_id = 'cloud-custodian'
-        factory = self.replay_flight_data('test_dataproc_clusters_query',
-                                          project_id=project_id)
-        p = self.load_policy(
-            {'name': 'dataproc_clusters',
-             'resource': 'gcp.dataproc-clusters'},
-            session_factory=factory)
-
-        resources = p.run()
-
-        self.assertEqual(len(resources), 1)
-        self.assertEqual(resources[0]['clusterName'], 'cluster-test')
